+++ Source/Boilerplate.Templates/Content/ApiTemplate/Controllers/CarsController.cs
@@ -117,11 +117,13 @@
         /// found.</returns>
 #if (Swagger)
         /// <response code="200">The car with the specified ID.</response>
+        /// <response code="304">The car has not changed since the date given in the If-Modified-Since HTTP header.</response>
         /// <response code="404">A car with the specified ID could not be found.</response>
 #endif
         [HttpGet("{carId}", Name = CarsControllerRoute.GetCar)]
         [HttpHead("{carId}")]
         [ProducesResponseType(typeof(Car), StatusCodes.Status200OK)]
+        [ProducesResponseType(typeof(void), StatusCodes.Status304NotModified)]
         [ProducesResponseType(typeof(void), StatusCodes.Status404NotFound)]
         public Task<IActionResult> Get(int carId) =>
             this.getCarCommand.Value.ExecuteAsync(carId);