+++ RELEASE NOTES.md
@@ -13,7 +13,8 @@ MVC 6 API
 - Added Prefix.io profiler option.
 - Upgraded NWebSec from 1.0.0 to 1.1.0.
 - Run health checks in parallel in StatusController.
-- 
+- Add Last-Modified and If-Modified-Since HTTP headers to return 304 Not Modified if a car has not changed.
+
 ### 1.5.0
 MVC 6 API
 - Added API Versioning support for Swagger.